GBy Sampling BM: Rename helper, clean up
