+++ test/engine/GroupByStrategyHelpers.h
